Move Oracles to own module
