+++ tests/active_learning_eval.py
@@ -1,7 +1,10 @@
-import unittest
-import pandas as pd
 import os
+import unittest
+
+import pandas as pd
+
 from active_learning.active_learning import *
+from active_learning.oracles import MockOracle
 from util.datastructures import MetaPath
